Added proper filters support for retreiving gateways
